% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{apodize_resolution}
\alias{apodize_resolution}
\title{Perform resolution enhancement}
\usage{
apodize_resolution(time, signal, R, tmax = NA, debug = FALSE)
}
\arguments{
\item{time}{Acquisition time in seconds.}

\item{signal}{Relative intensity values in the time domain.}

\item{R}{Decay rate paramter in Hz}

\item{tmax}{Location of Gaussian bell in seconds (estimated by default)}
}
\value{
Apodized and fourier transformed spectrum
}
\description{
Perform resolution enhancement
}
