% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{apodize_sensitivity}
\alias{apodize_sensitivity}
\title{Perform sensitivity enhancement}
\usage{
apodize_sensitivity(time, signal, R, debug = FALSE)
}
\arguments{
\item{time}{Acquisition time in seconds.}

\item{signal}{Relative intensity values in the time domain.}

\item{R}{Decay rate paramter in Hz}
}
\value{
Apodized and fourier transformed spectrum
}
\description{
Perform sensitivity enhancement
}
