% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{calculate_derivatives}
\alias{calculate_derivatives}
\title{Calculate derivatives using SG}
\usage{
calculate_derivatives(intensity, n)
}
\arguments{
\item{intensity}{Relative intensity values in the frequency domain}

\item{n}{Windows size (in points) for SG}

\item{smooth}{TRUE to take the absolute value and smooth the first derivative
of the real data and second derivative of the imaginary}
}
\value{
A data.frame containing the 1st and 2nd derivatives of the real and
        imaginary components (log-scaled)
}
\description{
Calculate derivatives using SG
}
