% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{calculate_time}
\alias{calculate_time}
\title{Convert chemical shift to time based on sweep frequency}
\usage{
calculate_time(chemical.shift, sf)
}
\arguments{
\item{chemical.shift}{Chemical shift in ppm}

\item{sf}{Sweep frequency in MHz}
}
\value{
A vector of time in seconds
}
\description{
Convert chemical shift to time based on sweep frequency
}
