% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{combine_pca}
\alias{combine_pca}
\title{Combine data using PCA}
\usage{
combine_pca(intensities, scale = FALSE)
}
\arguments{
\item{intensities}{A data.frame of intensity columns}

\item{scale}{TRUE to perform log scaling}
}
\value{
A vector of binary value specifying putative peaks
}
\description{
Combine data using PCA
}
