% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{count_peaks}
\alias{count_peaks}
\title{Count the number of peaks in each region}
\usage{
count_peaks(regions, peaks)
}
\arguments{
\item{regions}{A list of start/stop pairs indicating region bounds}

\item{peaks}{A vector of peak locations}
}
\value{
A vector of peak counts per region
}
\description{
Count the number of peaks in each region
}
