% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{dilate}
\alias{dilate}
\title{Wrapper around mmand::dilate to make it more convenient}
\usage{
dilate(logic, n)
}
\arguments{
\item{logic}{A vector of TRUE/FALSE values}

\item{n}{Dilation width}
}
\value{
A vector of TRUE/FALSE values
}
\description{
Wrapper around mmand::dilate to make it more convenient
}
