% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{estimate_decay}
\alias{estimate_decay}
\title{Estimate average decay rate}
\usage{
estimate_decay(time, signal, debug = FALSE)
}
\arguments{
\item{time}{Acquisition time in seconds.}

\item{signal}{Relative intensity values in the time domain.}

\item{sf}{Sweep frequency in MHz}
}
\value{
Average decay rate for the spectra.
}
\description{
Estimate average decay rate
}
