% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{estimate_n}
\alias{estimate_n}
\title{Estimate peak window size for SG filter}
\usage{
estimate_n(chemical.shift, width, height, sf)
}
\arguments{
\item{chemical.shift}{Chemical shift in ppm}

\item{width}{Narrowest peak width (at half height) in Hz}

\item{height}{Fraction of maximum peak height to include}

\item{sf}{Sweep frequency in MHz}
}
\value{
Minimum peak widths in points.
}
\description{
Estimate peak window size for SG filter
}
