% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmrfind.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks using ROI detection and resolution enhancement}
\usage{
find_peaks(
  chemical.shift,
  intensity,
  sf,
  peak.width = 0.8,
  dilation.ppm = 0.01,
  noise.threshold = 2,
  roi.noise.threshold = 4,
  step.threshold = 0.001,
  R = NA,
  tmax = NA,
  format = TRUE
)
}
\arguments{
\item{chemical.shift}{Sequence of chemical shift values in ppm}

\item{intensity}{Relative intensity values corresponding to each chemical
shift}

\item{sf}{Sweep frequency}

\item{peak.width}{Lower end of expected peak widths (in Hz, measured at half
height)}

\item{dilation.ppm}{Dilation factor for ROI detection (in ppm)}

\item{noise.threshold}{Multiple of noise standard deviation to set as cutoff
between signal and noise.}

\item{step.threshold}{Threshold step size during iterative peak detection
(smaller numbers increase computational time while
bigger numbers may miss peaks). Essentially, the noise
threshold is converted into a percentile value and this
value is incremented by the step.threshold until it
reaches 1. Reasonable options for this value are likely
to fall in the 0.001 to 0.1 range.}

\item{R}{Optional set of decay rates (in Hz) to assess for apodization
(calculated from the data by default)}

\item{tmax}{Optional set of tmax parameters to assess for apodization}

\item{format}{TRUE to output a list of peak locations in ppm, FALSE to output
raw representation of TRUE/FALSE values with TRUE values
specifying peaks}
}
\value{
A vector of TRUE/FALSE values or a list of peak locations in ppm
}
\description{
Find peaks using ROI detection and resolution enhancement
}
