% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnmrfind.R
\name{find_roi}
\alias{find_roi}
\title{Find regions of interest (ROI) using sensitivity enhancement}
\usage{
find_roi(
  chemical.shift,
  intensity,
  sf,
  peak.width = 1,
  dilation.ppm = 0.05,
  noise.threshold = 4,
  R = NULL,
  format = TRUE
)
}
\arguments{
\item{chemical.shift}{Sequence of chemical shift values in ppm}

\item{intensity}{Relative intensity values corresponding to each chemical
shift}

\item{sf}{Sweep frequency}

\item{peak.width}{Lower end of expected peak widths (in Hz, measured at half
height)}

\item{dilation.ppm}{Dilation factor for ROI detection (in ppm)}

\item{noise.threshold}{Multiple of noise standard deviation to set as cutoff
between signal and noise.}

\item{R}{Optional set of decay rates (in Hz) to assess for apodization
(calculated from the data by default)}

\item{format}{TRUE to output a list of vectors specifying ROI start and stop
(in ppm), FALSE to use indexes (1 to n), NULL to output raw
logic vector of TRUE/FALSE values where TRUE corresponds to an
ROI}
}
\value{
A vector of TRUE/FALSE values or a list of ROI start, stop vectors in
        ppm
}
\description{
Find regions of interest (ROI) using sensitivity enhancement
}
