% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{group_peaks}
\alias{group_peaks}
\title{Group peaks based on region}
\usage{
group_peaks(regions, peaks)
}
\arguments{
\item{regions}{A list of start/stop pairs indicating region bounds}

\item{peaks}{A vector of peak locations}
}
\value{
A vector of peak counts per region
}
\description{
Group peaks based on region
}
