% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{list_peaks}
\alias{list_peaks}
\title{Convert a series of TRUE/FALSE values into peak positions based on local
maxima}
\usage{
list_peaks(logic, reference, values = NULL)
}
\arguments{
\item{logic}{A vector of TRUE/FALSE values}

\item{reference}{A set of values to check for local maxima}

\item{values}{Optional vector of values to output instead of indexes}
}
\value{
A vector of peak positions (either as indexes or values)
}
\description{
Convert a series of TRUE/FALSE values into peak positions based on local
maxima
}
