% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{list_regions}
\alias{list_regions}
\title{Convert a series of TRUE/FALSE values into start/stop indexes or values}
\usage{
list_regions(logic, values = NULL)
}
\arguments{
\item{logic}{A vector of TRUE/FALSE values}

\item{values}{Optional vector of values to output instead of indexes}
}
\value{
A list of start/stop pairs (either as indexes or values)
}
\description{
Convert a series of TRUE/FALSE values into start/stop indexes or values
}
