% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr.spectra}
\alias{nmr.spectra}
\title{A list of data frames with example 1H NMR data.}
\format{
List.
}
\usage{
nmr.spectra
}
\description{
There are 10 data frames total, one for alanine, arginine, cysteine, folic
acid, fructose, pyridoxine, serine, sucrose, valeric acid, and valine. Each
data.frame has two columns: "chemical.shift" (numeric) and "intensity"
(complex). Raw 1H NMR data was downloaded from HMD and phase corrected using
TopSpin.
}
\details{
HMDB citation: Wishart DS, Feunang YD, Marcu A, Guo AC, Liang K, et al.,
HMDB 4.0 — The Human Metabolome Database for 2018. Nucleic Acids Res. 2018.
Jan 4;46(D1):D608-17. 29140435
}
\keyword{datasets}
