% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{open}
\alias{open}
\title{Wrapper around mmand::opening to make it more convenient}
\usage{
open(logic, n)
}
\arguments{
\item{logic}{A vector of TRUE/FALSE values}

\item{n}{Opening width}
}
\value{
A vector of TRUE/FALSE values
}
\description{
Wrapper around mmand::opening to make it more convenient
}
