% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot overlapping NMR spectra using plotly}
\usage{
plot_spectra(chemical.shift, intensities, scale = FALSE)
}
\arguments{
\item{chemical.shift}{Sequence of chemical shift values in ppm.}

\item{intensity}{Complex intensity values corresponding to each chemical
shift.}
}
\description{
Plot overlapping NMR spectra using plotly
}
