% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot NMR spectrum using plotly}
\usage{
plot_spectrum(
  chemical.shift,
  intensity,
  imaginary = FALSE,
  peaks = c(),
  regions = list()
)
}
\arguments{
\item{chemical.shift}{Sequence of chemical shift values in ppm.}

\item{intensity}{Complex intensity values corresponding to each chemical
shift.}
}
\description{
Plot NMR spectrum using plotly
}
