% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{trim}
\alias{trim}
\title{Set n values on either side of vector to FALSE}
\usage{
trim(logic, n)
}
\arguments{
\item{logic}{A vector of TRUE/FALSE values}

\item{n}{Number of values to make FALSE on either side of vector}
}
\value{
A vector of TRUE/FALSE values
}
\description{
Set n values on either side of vector to FALSE
}
